+++ cardano-tracer/bench/cardano-tracer-bench.hs
@@ -4,7 +4,7 @@ import           Control.Concurrent.Extra (newLock)
 import           Control.Concurrent.STM.TVar (newTVarIO)
 import           Criterion.Main
 import qualified Data.List.NonEmpty as NE
-import           Data.Time.Clock (getCurrentTime)
+import           Data.Time.Clock (UTCTime, getCurrentTime)
 import           System.Directory (getTemporaryDirectory, removePathForcibly)
 import           System.FilePath ((</>))
 
@@ -19,6 +19,8 @@ import           Cardano.Tracer.Environment
 import           Cardano.Tracer.Handlers.Logs.TraceObjects
 import           Cardano.Tracer.MetaTrace
 import           Cardano.Tracer.Types
+
+import qualified Criterion.Types as Criterion
 
 main :: IO ()
 main = do
@@ -76,22 +78,24 @@ main = do
 
   removePathForcibly root
 
-  defaultMain
+  defaultMainWith defaultConfig { Criterion.verbosity = Criterion.Verbose }
     [ bgroup "cardano-tracer"
-      [ -- 10 'TraceObject's per request.
-        bench "Handle TraceObjects LOG,  10"   $ whnfIO $ traceObjectsHandler te1 nId to10
-      , bench "Handle TraceObjects JSON, 10"   $ whnfIO $ traceObjectsHandler te2 nId to10
+      [  -- 10 'TraceObject's per request.
+        bench "Handle TraceObjects LOG,  10"   $ whnfIO $ beforeProgramStops do traceObjectsHandler te1 nId to10
+      , bench "Handle TraceObjects JSON, 10"   $ whnfIO $ beforeProgramStops do traceObjectsHandler te2 nId to10
         -- 100 'TraceObject's per request.
-      , bench "Handle TraceObjects LOG,  100"  $ whnfIO $ traceObjectsHandler te1 nId to100
-      , bench "Handle TraceObjects JSON, 100"  $ whnfIO $ traceObjectsHandler te2 nId to100
+      , bench "Handle TraceObjects LOG,  100"  $ whnfIO $ beforeProgramStops do traceObjectsHandler te1 nId to100
+      , bench "Handle TraceObjects JSON, 100"  $ whnfIO $ beforeProgramStops do traceObjectsHandler te2 nId to100
         -- 1000 'TraceObject's per request.
-      , bench "Handle TraceObjects LOG,  1000" $ whnfIO $ traceObjectsHandler te1 nId to1000
-      , bench "Handle TraceObjects JSON, 1000" $ whnfIO $ traceObjectsHandler te2 nId to1000
+      , bench "Handle TraceObjects LOG,  1000" $ whnfIO $ beforeProgramStops do traceObjectsHandler te1 nId to1000
+      , bench "Handle TraceObjects JSON, 1000" $ whnfIO $ beforeProgramStops do traceObjectsHandler te2 nId to1000
       ]
     ]
  where
+  nId :: NodeId
   nId = NodeId "run-user-1000-cardano-tracer-demo.sock@0"
 
+  mkConfig :: FilePath -> LogFormat -> TracerConfig
   mkConfig root format = TracerConfig
     { networkMagic   = 764824073
     , network        = AcceptAt (LocalSocket "")
@@ -105,10 +109,13 @@ main = do
     , verbosity      = Nothing
     , metricsComp    = Nothing
     , hasForwarding  = Nothing
+    , resourceFreq   = Nothing
     }
 
+  generate :: Int -> IO [TraceObject]
   generate num = replicate num . mkTraceObject <$> getCurrentTime
 
+  mkTraceObject :: UTCTime -> TraceObject
   mkTraceObject now = TraceObject
     { toHuman     = Just "Human Message About Some Important Information From The Cardano Node"
     , toMachine   = "{\"msg\": \"forMachine Important Message\"}"