+++ nix/workbench/manifest.sh
@@ -42,7 +42,7 @@ case "${op}" in
         def is_from_chap:
           ."pkg-src".repo?.uri?
           | strings
-          | . == "https://input-output-hk.github.io/cardano-haskell-packages" or startswith("file:" + $chappath)
+          | . == "https://chap.intersectmbo.org/" or startswith("file:" + $chappath)
           ;
 
         def extract_package_data:
@@ -171,7 +171,22 @@ case "${op}" in
         local json=${1:?$usage}
 
         jq '
-        include "lib";
+        def color:
+          { red:     "\u001b[31m"
+          , green:   "\u001b[32m"
+          , yellow:  "\u001b[33m"
+          , blue:    "\u001b[34m"
+          , magenta: "\u001b[35m"
+          , cyan:    "\u001b[36m"
+          , white:   "\u001b[37m"
+          , off:     "\u001b[39m"
+          };
+
+        def colorly($col; $x):
+          "\(color[$col])\($x)\(color["off"])";
+
+        def drop_nulls:
+          map(select(. != null));
 
         def unwords: join(" ");
         def unlines: join("\n");