+++ nix/workbench/service/nodes.nix
@@ -1,6 +1,7 @@
 { pkgs
 , runJq
 , runWorkbench
+, jsonFilePretty
 
 ## The cardano-node config used as baseline:
 , baseNodeConfig
@@ -13,9 +14,14 @@
 }:
 
 with pkgs.lib;
-with (import ../lib.nix pkgs.lib);
 
 let
+  readJSONMay = fp:
+    let fv = __tryEval (__readFile fp);
+    in if fv.success
+       then __fromJSON fv.value
+       else {};
+
   profileName = profile.name;
 
   eras = [ ## This defines the order of eras -- which is important.
@@ -201,10 +207,10 @@ let
 
       config = {
         value = service.nodeConfig;
-        JSON  = runJq "node-config-${name + modeIdSuffix}.json"
-                  ''--null-input --sort-keys
-                    --argjson x '${__toJSON service.nodeConfig}'
-                  '' "$x";
+        JSON  = jsonFilePretty
+                  "node-config-${name + modeIdSuffix}.json"
+                  (__toJSON service.nodeConfig)
+        ;
       };
 
       topology =