+++ nix/workbench/shell.nix
@@ -19,7 +19,7 @@ with lib;
 let
 
     # recover CHaP location from cardano's project
-    chap = project.args.inputMap."https://input-output-hk.github.io/cardano-haskell-packages";
+    chap = project.args.inputMap."https://chap.intersectmbo.org/";
 
     # build plan as computed by nix
     nixPlanJson = project.plan-nix.json;