+++ cardano-submit-api/CHANGELOG.md
@@ -17,8 +17,8 @@
 
 ## 8.0.0 -- May 2023
 
-- [Add tx_submit_fail_count metric](https://github.com/input-output-hk/cardano-node/pull/4566)
-- [Configurable metrics port in submit-api](https://github.com/input-output-hk/cardano-node/pull/4281)
+- [Add tx_submit_fail_count metric](https://github.com/intersectmbo/cardano-node/pull/4566)
+- [Configurable metrics port in submit-api](https://github.com/intersectmbo/cardano-node/pull/4281)
 
 ## 1.35.3 -- July 2022