+++ cardano-tracer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## 0.2.2
+
+* Add resource tracing for `cardano-tracer`.
+* Add config option: `resourceFreq`.
+* Fix benchmark suite.
+
 ## 0.2.1
 
 * Updated to `ouroboros-network-0.10`